JAMES-1968 Separate Scripts operations and quota operation on Cassandra Sieve storage
